+++ plugins/a11yheading/lang/en.js
@@ -12,8 +12,8 @@ CKEDITOR.plugins.setLang( 'a11yheading', 'en', {
   level_h5:       'H5 – Subsection title',
   level_h6:       'H6 – Subsection title',
   format_p:       'Normal',
-  format_pre:     'Computer code',
-  format_address: 'Address',
+  format_pre:     'Preformatted text',
+  format_address: 'Address line',
   format_div:     'Normal (DIV)',
   helpLabel:      'Help'
 } );